feat: 기존 SchedulerService 위치 이동 및 식물 API 영속화 작업 추가 #55
